# Make Time initialization faster