This makes most `Time` initialization calls a bit faster, and `Time.utc` **much** faster.

The main improvement here is storing the two most commonly used fstrings used by this file, `"UTC"` and `""`, in a static global. This way we don't need to fetch them from the fstring table each time we construct a new time. This makes `Time.utc` nearly twice as fast, and most other methods slightly faster.

## Benchmark

```
require "benchmark/ips"

Benchmark.ips do |x|
  x.report "Time.utc" do
    Time.utc(2020, 1, 1, 0, 0, 0, 0)
  end

  x.report "Time.at" do
    Time.at(1577836800)
  end
end
```

**Before:**
 
```
            Time.utc      2.273M (± 0.6%) i/s -     11.436M in   5.031174s
          Time.local      1.112M (± 0.2%) i/s -      5.650M in   5.079622s
            Time.new      2.196M (± 0.6%) i/s -     11.094M in   5.051065s
            Time.now      5.724M (± 0.8%) i/s -     28.643M in   5.004617s
             Time.at      9.394M (± 0.6%) i/s -     47.296M in   5.034809s
```

**After**

```
            Time.utc      4.362M (± 0.2%) i/s -     22.064M in   5.058701s
          Time.local      1.326M (± 0.4%) i/s -      6.679M in   5.035385s
            Time.new      3.028M (± 0.4%) i/s -     15.140M in   5.000674s
            Time.now      5.785M (± 0.2%) i/s -     29.271M in   5.059968s
             Time.at      8.902M (± 0.1%) i/s -     44.589M in   5.009060s
```